+++ lib/src/utilities/_app_config.dart
@@ -1,5 +1,7 @@
 class AppConfig {
   static String url = 'https://backend.connect-x.tech';
+  // static String url = 'https://connect-x-gun.connect-x.tech';
+
   static String token =
       r"Vmh;{v]M[pJqJMC2-bj@R<ZYeMred0BT,aHL*QYhPD5/wj%/^hVz4J=Ts$,+%:SS";